unexe/requirements.txt
psycopg2>=2.2,<3
#Django>=1.5,<1.6
#django-registration>=1.0,<2
#-e hg+https://bitbucket.org/fcurella/django-profiles@5482ded#egg=django-profiles
#djangorestframework>=2.3,<3
#South>=0.8
#django-notify>=1.1,<2
#django-ajax-selects>=1.2.5,<2
#Markdown>=2.1,<3
#django-simple-captcha>=0.4
pthelma>=0.2.1
#django-appconf>=0.6
gdal>=1.6
django-tables2>=0.14
